Addressed a few issues. Added comment removal to original API. Added Obsolete tags to InsertPatch and PrefabExtensionInsertPatch. Removed superfluous V2 CustomPatch.
